+++ atomicshop/basics/bytes_arrays.py
@@ -51,3 +51,63 @@ def convert_sequence_of_bytes_to_sequence_of_strings(byte_sequence: bytes) -> li
         # single_byte_string = byte_sequence[byte_index:byte_index+1]
 
     return result
+
+
+def find_position(file_path: str, target: bytes, chunk_size: int = None, starting_position: int = 0) -> int:
+    """
+    Find position of the target bytes string in the file.
+
+    :param file_path: string, path to file.
+    :param target: bytes, target bytes string.
+    :param chunk_size: integer, chunk size  in bytes.
+    :param starting_position: integer, starting position in bytes. You can specify the starting seeking point
+        in the file.
+    :return:
+    """
+
+    if not chunk_size:
+        chunk_size = len(target)
+
+    # chunk_size = 4096  # You can adjust the chunk size based on your needs
+
+    # Overlap between chunks to ensure target isn't split between chunks
+    overlap_size = len(target) - 1
+
+    with open(file_path, 'rb') as file:
+        # Move the file cursor to the starting position.
+        file.seek(starting_position)
+        # Update the position variable to match the starting position.
+        position = starting_position
+        chunk = file.read(chunk_size)
+
+        while chunk:
+            index = chunk.find(target)
+            if index != -1:
+                # Return the absolute position of the target in the file
+                return position + index
+
+            # Move the file cursor back by the overlap size to ensure target isn't split between chunks
+            file.seek(position + chunk_size - overlap_size)
+            # Get the current position of the cursor in the file.
+            position = file.tell()
+            chunk = file.read(chunk_size)
+
+    # Return -1 if the target is not found
+    return -1
+
+
+def read_bytes_from_position(file_path: str, starting_position: int, num_bytes: int) -> bytes:
+    """
+    Read bytes from specified position in the file.
+    :param file_path: string, path to file.
+    :param starting_position: integer, starting position in bytes.
+    :param num_bytes: integer, number of bytes to read.
+    :return: bytes.
+    """
+
+    with open(file_path, 'rb') as file:
+        # Move the file cursor to the specified position.
+        file.seek(starting_position)
+        # Read the specified number of bytes.
+        data = file.read(num_bytes)
+    return data